+++ src/main/java/com/bora/service/openbank/OpenBankingService.java
@@ -15,6 +15,14 @@ import com.bora.domain.openbank.account.AccountSearchRequestVO;
 import com.bora.domain.openbank.account.AccountSearchResponseVO;
 import com.bora.domain.openbank.account.AccountTranRequestVO;
 import com.bora.domain.openbank.account.AccountTranResponseVO;
+import com.bora.domain.openbank.card.CardInfoRequestVO;
+import com.bora.domain.openbank.card.CardInfoResponseVO;
+import com.bora.domain.openbank.card.CardListRequestVO;
+import com.bora.domain.openbank.card.CardListResponseVO;
+import com.bora.domain.openbank.card.bill.CardBillsRequestVO;
+import com.bora.domain.openbank.card.bill.CardBillsResponseVO;
+import com.bora.domain.openbank.card.bill.CardDetailBillsRequestVO;
+import com.bora.domain.openbank.card.bill.CardDetailBillsResponseVO;
 import com.bora.domain.openbank.tran.deposit.TranDepositRequestVO;
 import com.bora.domain.openbank.tran.deposit.TranDepositResponseVO;
 import com.bora.domain.openbank.tran.result.TranResultRequestVO;
@@ -44,11 +52,6 @@ public class OpenBankingService {
 		return openBankingApiClient.findAccount(accountSearchRequestVO);
 	}
 	
-	//등록계좌 해지
-	public AccountCancelResponseVO cancelAccount(AccountCancelRequestVO accountCancelRequestVO) {
-		return openBankingApiClient.cancelAccount(accountCancelRequestVO);
-	}
-	
 	//잔액조회
 	public AccountBalanceResponseVO findAccountBalance(AccountBalanceRequestVO accountBalanceRequestVO) {
 		return openBankingApiClient.findAccountBalance(accountBalanceRequestVO);
@@ -59,20 +62,26 @@ public class OpenBankingService {
 		return openBankingApiClient.findAccountTran(accountTranRequestVO);
 	}
 
-	//출금이체
-	public TranWithdrawResponseVO findTranWithdraw(TranWithdrawRequestVO tranWithdrawRequestVO) {
-		return openBankingApiClient.findTranWithdraw(tranWithdrawRequestVO);
+	// 카드기본정보 조회
+	public CardInfoResponseVO infoCard(CardInfoRequestVO cardInfoRequestVO) throws Exception {
+		return openBankingApiClient.infoCard(cardInfoRequestVO);
 	}
 	
-	//입금이체
-	public TranDepositResponseVO findTranDeposit(TranDepositRequestVO tranDepositRequestVO) {
-		return openBankingApiClient.findTranDeposit(tranDepositRequestVO);
+	// 카드목록 조회
+	public CardListResponseVO listCard(CardListRequestVO cardListRequestVO) throws Exception {
+		return openBankingApiClient.listCard(cardListRequestVO);
 	}
-
-	//이체결과조회
-	public TranResultResponseVO findTranResult(TranResultRequestVO tranResultRequestVO) {
-		return openBankingApiClient.findTranResult(tranResultRequestVO);
+	
+	// 카드청구기본정보 조회
+	public CardBillsResponseVO billsCard(CardBillsRequestVO cardBillsRequestVO) throws Exception {
+		return openBankingApiClient.billsCard(cardBillsRequestVO);
 	}
+	
+	// 카드청구상세정보 조회
+	public CardDetailBillsResponseVO detailBillsCard(CardDetailBillsRequestVO cardDetailBillsRequestVO) throws Exception {
+		return openBankingApiClient.detailBillsCard(cardDetailBillsRequestVO);
+	}
+	
 	
 	
 }
