--- src/main/webapp/WEB-INF/views/card/card_user_info.jsp
@@ -1,14 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="UTF-8">
-<title>Insert title here</title>
-</head>
-<body>
-    <h1>사용자 정보조회 결과</h1>
-	<h3>고객번호 : ${userInfo.user_seq_no }</h3>
-	<h3>고객CI값 : ${userInfo.user_ci }</h3>
-</body>
-</html>
